Moved getOrGenerateSiteUuid into settings-utils
